+++ packages/cli/CHANGELOG.md
@@ -2,5 +2,8 @@
 
 ## 0.5.0
 
-* Configure single CLI entrypoint.
+* Make language-related functionality "multi-lingual", i.e. a serialization chunk can contain multiple `Language`s.
+  * (Languages will be sorted by name.)
+* Improve extraction functionality: only catch JSON-parsing exceptions.
+* Configure single entrypoint named `lionweb-cli`.