+++ packages/utilities/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 0.5.0
 
+* Extend textual syntax for "multi-linguality".
+* Implement textual syntax for `Annotation`s.
 * _Breaking change:_ Change ID generation to hashing (`hasher` function).
-* (Make small fix in `asText` function.)
+* (Make small, cosmetic fix in `asText` function.)
 * Add a function `tsTypesForLanguage` that generates TypeScript types.
 * Receive `asText` and ID checking from `@lionweb/core`.